pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "flowpilot-challenge"
version = "0.1.0"
description = "Challenge scaffold"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Zoner" }]
dependencies = [
    "typer[all]>=0.12",
    "pydantic>=2.7",
    "pydantic-settings>=2.2",
    "python-dotenv>=1.0",
    "rich>=13.7",
    "structlog>=24.1",
    "orjson>=3.10",
    "datasets>=2.18",
    "transformers>=4.40",
    "sentence-transformers>=2.6",
    "accelerate>=0.28",
    "faiss-cpu>=1.7.4",
    "langchain>=0.2",
    "langchain-community>=0.2",
    "langchain-openai>=0.1",
    "langchainhub>=0.1",
    "chromadb>=0.5",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "psycopg2-binary>=2.9",
    "pgvector>=0.3.0",
    "langgraph>=0.0.10",
    "tavily-python>=0.3.0",
    "langgraph-cli>=0.4.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "mypy>=1.10",
    "ruff>=0.4",
    "types-requests",
]

[project.scripts]
agentic-rag = "agentic_rag.cli:app"

[tool.ruff]
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "RUF"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
addopts = "-ra --showlocals"
pythonpath = ["src", "."]
markers = [
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
branch = true
source = ["agentic_rag"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]